add randomization of tokens
- add player starting room randomization
- add alien placement room randomization
- add docstring to game_play function
- refactored for modularization
